fixed directional light
it was a problem with the light classes, too many people tried to
refactor them separately and we all got confused .... so this is a fix
for all that nonsense. took "position" out of the light struct and just
used the transform node instead of setting position to the transform
etc. etc. blah blah blah