fixed mouse stuff
The way mouse input was handled earlier went against GLFW design
conventions, so I changed it to match the suggested use: the input mode
is set to GLFW_CURSOR_DISABLED, which makes the mouse position
unbounded. Also added a thing which tracks bounded mouse position, since
we probably need that.