fixed an incredibly confusing memory error
It makes sense that trying to delete that shader would mess things up
since it's set to autorelease and the last thing to use it would also be
calling delete. Instead of breaking when trying to delete the shader for
a second time, the application continued closing successfully until it
reached the glfwTerminate, at which point it broke while trying to
destroy the window. idk why